Refactor dm composer to look up an existing thread if preloaded with a user id
